--- core/data/src/main/java/com/client/ai/MediaPipeEmbeddings.kt
@@ -1,13 +1,0 @@
-package com.client.ai
-
-import com.client.model.JobSimilarity
-
-interface MediaPipeEmbeddings {
-
-    suspend fun setUpMLModel()
-
-    suspend fun getSimilarities(
-        mainSentence: String,
-        sentences: List<String>
-    ): List<JobSimilarity>
-}
