--- feature/account/data/src/main/java/com/client/account/data/MyClass.kt
@@ -1,4 +1,0 @@
-package com.client.account.data
-
-class MyClass {
-}
