--- feature/account/domain/src/main/java/com/client/account/domain/MyClass.kt
@@ -1,4 +1,0 @@
-package com.client.account.domain
-
-class MyClass {
-}
