+++ feature/matching/data/src/main/java/com/client/account/data/MediaPipeEmbeddings.kt
@@ -1,0 +1,13 @@
+package com.client.account.data
+
+import com.client.model.JobSimilarity
+
+interface MediaPipeEmbeddings {
+
+    suspend fun setUpMLModel()
+
+    suspend fun getSimilarities(
+        mainSentence: String,
+        sentences: List<String>
+    ): List<JobSimilarity>
+}
