+++ feature/profile/data/build.gradle.kts
@@ -6,7 +6,7 @@ plugins {
 }
 
 android {
-    namespace = "com.client.employ.account.data"
+    namespace = "com.client.employ.profile.data"
 
     buildFeatures {
         buildConfig = true
