+++ feature/profile/domain/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    id("java-library")
+    alias(libs.plugins.kotlin.jvm)
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_17
+    targetCompatibility = JavaVersion.VERSION_17
+}
